% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_embeddings.R
\name{autoplot.rbert_pca}
\alias{autoplot.rbert_pca}
\title{Plot PCA}
\usage{
\method{autoplot}{rbert_pca}(pca_df, color_field = NULL)
}
\arguments{
\item{pca_df}{A tbl_df of pca vectors with class "rbert_pca"; output from a
call to \code{do_pca}.}

\item{color_field}{Character scalar; optional column name to assign to color
aesthetic in the plot.}
}
\value{
A ggplot2 plot of the embedding vectors projected onto two principal
  axes.
}
\description{
Given the output of \code{do_pca}, make the plot object.
}
\examples{
\dontrun{
# assuming something like the following has been run:
# feats <- RBERT::extract_features(...) # See RBERT documentation
# Then:
feats$output \%>\%
    filter_layer_embeddings(layer_indices = 12L) \%>\%
    keep_tokens("[CLS]") \%>\%
    do_pca() \%>\%
    autoplot()
}
}
