% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_attention_matrix}
\alias{convert_attention_matrix}
\title{Restructure Attention Data}
\usage{
convert_attention_matrix(att_ten_list)
}
\arguments{
\item{att_ten_list}{List of numerical tensors of attention weights. Expected
structure: \code{[layer_index][head_index, token_from_index,
token_to_index]}}
}
\value{
The input attention weights, restructured as nested lists.
}
\description{
Converts a list of attention tensors to the nested list format required
downstream.
}
\keyword{internal}
