% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_embeddings.R
\name{display_pca}
\alias{display_pca}
\title{Display PCA of Embeddings}
\usage{
display_pca(
  embedding_df,
  color_field = NULL,
  disambiguate_tokens = TRUE,
  hide = NULL
)
}
\arguments{
\item{embedding_df}{A tbl_df of embedding vectors; the output of
\code{\link[RBERT]{extract_features}}.}

\item{color_field}{Character scalar; optional column name to assign to color
aesthetic in the plot.}

\item{disambiguate_tokens}{Logical; whether to append example and token
index to the literal token for display purposes.}

\item{hide}{Optional logical vector of same length as
\code{nrow(embedding_df)}; if present, only rows corresponding to TRUE will
be displayed on the plot (all rows will still be used in the computation of
the plot). This makes it easy to selectively display points without
redefining the principal axes.}

\item{class}{Character vector; which tokens to keep.}
}
\value{
A ggplot2 plot of the embedding vectors projected onto two principal
  axes.
}
\description{
Display a 2D PCA plot of a collection of embedding vectors.
}
\examples{
\dontrun{
# assuming something like the following has been run:
# feats <- RBERT::extract_features(...) # See RBERT documentation
# Then:
feats$output %>%
    filter_layer_embeddings(layer_indices = 12L) %>%
    keep_tokens("[CLS]") %>%
    display_pca()
}
}
