% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_embeddings.R
\name{do_pca}
\alias{do_pca}
\title{Do PCA on Embeddings}
\usage{
do_pca(
  embedding_df,
  project_vectors = embedding_df,
  color_field = NULL,
  disambiguate_tokens = TRUE
)
}
\arguments{
\item{embedding_df}{A tbl_df of embedding vectors; from the output of
\code{\link[RBERT]{extract_features}}.}

\item{project_vectors}{A tbl_df of embedding vectors to be used for
calculating the PCA projection matrix. Defaults to \code{embedding_df}.
This makes it possible to more consistently select the PCA "perspective",
even as the set of vectors may change.}

\item{color_field}{Character scalar; optional column name to assign to color
aesthetic in the plot.}

\item{disambiguate_tokens}{Logical; whether to append example and token
index to the literal token for display purposes.}
}
\value{
A tbl_df of the embedding vectors projected onto two principal
  axes.
}
\description{
Calculate the 2D PCA for a collection of embedding vectors, in preparation
for plotting.
}
\examples{
\dontrun{
# assuming something like the following has been run:
# feats <- RBERT::extract_features(...) # See RBERT documentation
# Then:
pca_df <- feats$output \%>\%
    filter_layer_embeddings(layer_indices = 12L) \%>\%
    keep_tokens("[CLS]") \%>\%
    do_pca()
}
}
