% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.enlist_attention}
\alias{.enlist_attention}
\title{Convert an Attention df to a List}
\usage{
.enlist_attention(attention_tibble, top_segment_index, bottom_segment_index)
}
\arguments{
\item{attention_tibble}{A tibble of attention, with minimal columns
layer_index, head_index, token_index, attention_token_index, and
attention_weight.}

\item{top_segment_index}{Integer; the index of the first/top segment for this
set. Can be a vector (ie, \code{c(1, 2)} for "all").}

\item{bottom_segment_index}{Integer; the index of the second/bottom segment
for this set. Can be a vector (ie, \code{c(1, 2)} for "all").}
}
\value{
A list as expected by \code{\link{.format_attention}}.
}
\description{
Convert an Attention df to a List
}
\keyword{internal}
