% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_attention}
\alias{.format_attention}
\title{Format BERT Attention Weights for Visualizing}
\usage{
.format_attention(attention, sequence_index = 1)
}
\arguments{
\item{attention}{The "attention" component of the output from
\code{\link[RBERT]{extract_features}(..., features = "attention")}.}

\item{sequence_index}{Integer; which example sequence from the input to
visualize.}
}
\value{
an object structured appropriately for input (as eventual json) to
  the javascript visualization code.
}
\description{
Given data from an RBERT model, format the attention weights matrices in the
structure expected by the javascript visualization code.
}
\keyword{internal}
