% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_attention_tokens}
\alias{.get_attention_tokens}
\title{Pull the Designated Tokens from the Attention tibble}
\usage{
.get_attention_tokens(attention, this_segment_index)
}
\arguments{
\item{attention}{The attention component from
\code{\link[RBERT]{extract_features}(..., features = "attention")}.}

\item{this_segment_index}{The segment to pull.}
}
\value{
A character vector of tokens.
}
\description{
Pull the Designated Tokens from the Attention tibble
}
\keyword{internal}
