% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_embeddings.R
\name{extract_vectors_df}
\alias{extract_vectors_df}
\title{Extract Embedding Vectors}
\usage{
extract_vectors_df(layer_outputs)
}
\arguments{
\item{layer_outputs}{The \code{layer_outputs} component of
\code{\link[RBERT]{extract_features}} output.}
}
\value{
The embedding vector components as a tbl_df, for all tokens and all
  layers.
}
\description{
Extract the embedding vector values from output of
\code{\link[RBERT]{extract_features}}. The resulting tbl_df will typically have a
large number of columns (> 768), so it will be rather slow to \code{\link{View}}.
Consider using \code{\link[dplyr]{glimpse}} or \code{\link[useful]{corner}}
if you just want to peek at the values.
}
\examples{
\dontrun{
# assuming something like the following has been run:
# feats <- RBERT::extract_features(...) # See RBERT documentation
# Then:
embeddings <- extract_vectors_df(feats$layer_outputs)
}
}
