% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_embeddings.R
\name{filter_layer_embeddings}
\alias{filter_layer_embeddings}
\title{Filter and Combine Embedding Vectors}
\usage{
filter_layer_embeddings(embedding_df, layer_indices = 12L, sum_fun = mean)
}
\arguments{
\item{embedding_df}{A tbl_df of embedding vectors; the "output" component of
\code{\link[RBERT]{extract_features}}.}

\item{layer_indices}{Integer vector; which layers embeddings to keep.}

\item{sum_fun}{A summarizing function to apply to the embedding vector
components of the retained layers. Will be passed as the \code{.funs}
argument to \code{\link[dplyr]{summarize_at}}. By default,
\code{\link{mean}} is used, but other functions are possible. If this
parameter is given a NULL value, no summarizing will be done, and only the
filtered tbl_df will be returned.}
}
\value{
The input tbl_df of embedding vectors, with the specified filtering
  and summarizing applied.
}
\description{
Applies a filter + summarize to the given table of embeddings. Can be used to
filter to specific layer outputs, or to average the outputs from multiple
layers.
}
\examples{
\dontrun{
# assuming something like the following has been run:
# feats <- RBERT::extract_features(...) # See RBERT documentation
# Then:
embeddings <- feats$output
embeddings_layer1 <- embeddings \%>\%
    filter_layer_embeddings(layer_indices = 1L)
# To keep the layer_index column, use sum_fun = NULL:
embeddings_layer1b <- embeddings \%>\%
    filter_layer_embeddings(layer_indices = 1L, sum_fun = NULL)
# Take the component-wise mean of the embeddings across layers 10-12.
embeddings_layer_mean <- embeddings \%>\%
    filter_layer_embeddings(layer_indices = 10:12)
# Take the component-wise maximum of the embeddings across layers 8-12.
# (Why would we do this? I don't know.)
embeddings_layer_max <- embeddings \%>\%
    filter_layer_embeddings(layer_indices = 8:12, sum_fun = max)
}
}
