% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_attention}
\alias{format_attention}
\title{Format BERT Attention Weights for Visualizing}
\usage{
format_attention(att_mat, seq_num = 1)
}
\arguments{
\item{att_mat}{The "attention_probs" component of the output from
\code{RBERT::extract_features()} (details to come).}

\item{seq_num}{Integer; which example sequence from the input to visualize.}
}
\value{
an object structured appropriately for input (as eventual json) to
  the javascript visualization code.
}
\description{
Given data from an RBERT model, format the attention weights matrices in the
structure expected by the javascript visualization code.
}
\examples{
\dontrun{
# assuming something like the following has been run:
# feats <- RBERT::extract_features(...) # See RBERT documentation
# Then:
attn <- format_attention(feats$attention_probs)
}
}
