% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\arguments{
\item{lhs, rhs}{A vector of fields or a tibble of fields and values, and a
function to apply to them.}
}
\description{
The pipe function, \code{\%>\%}, allows you to turn function composition into
a series of imperative statements.
}
