% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_sequences}
\alias{split_sequences}
\title{Split a Token Sequence}
\usage{
split_sequences(token_list)
}
\arguments{
\item{token_list}{List of tokens.}
}
\value{
Two lists of tokens, divided on the first \code{"[SEP]"} token.
}
\description{
Splits a sequence of tokens on the first \code{"[SEP]"} token. Everything
up to and including the first \code{"[SEP]"} is in the first segment,
everything following is in the second.
}
\keyword{internal}
