% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_attention.R
\name{visualize_attention}
\alias{visualize_attention}
\title{Visualize BERT Attention Weights}
\usage{
visualize_attention(
  attention,
  sequence_index = 1,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{attention}{The "attention" component of the output from
\code{\link[RBERT]{extract_features}(..., features = "attention")}.}

\item{sequence_index}{Integer; which example sequence from the input to
visualize.}

\item{height, width}{Characters; height and width of the htmlwidget
specified in any valid \code{CSS} size unit.}

\item{elementId}{Character; a valid \code{CSS} element id.}
}
\value{
the output from \code{htmlwidgets::createWidget}.
}
\description{
Given data from an RBERT model, display an interactive visualization of the
attention weights.
}
\examples{
\dontrun{
# assuming something like the following has been run:
# feats <- RBERT::extract_features(...) # See RBERT documentation
# Then:
visualize_attention(feats$attention)
}
}
