% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_attention.R
\name{visualize_attention_html}
\alias{visualize_attention_html}
\title{Customize Visualizer html}
\usage{
visualize_attention_html(id, style, class, ...)
}
\arguments{
\item{id}{Parameter used by \code{htmlwidgets}.}

\item{style}{Parameter used by \code{htmlwidgets}.}

\item{class}{Parameter used by \code{htmlwidgets}.}

\item{...}{Other parameters used by \code{htmlwidgets}.}
}
\value{
The output from htmltools::tags$span(). I think this should be
  properly formatted html, as a string.
}
\description{
"This function is looked up within the package implementing the widget by the
convention widgetname_html so it need not be formally exported from your
package or otherwise registered with htmlwidgets."
}
\details{
You shouldn't need to call this function directly.
}
\keyword{internal}
