% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_attention.R
\name{visualize_attention}
\alias{visualize_attention}
\title{Visualize BERT Attention Weights}
\usage{
visualize_attention(attn_obj, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{attn_obj}{Output from \code{format_attention()} (details to come).}

\item{height, width}{Characters; height and width of the htmlwidget
specified in any valid \code{CSS} size unit.}

\item{elementId}{Character; a valid \code{CSS} element id.}
}
\value{
the output from \code{htmlwidgets::createWidget}.
}
\description{
Given data from an RBERT model, display an interactive visualization of the
attention weights.
}
\examples{
\dontrun{
# assuming something like the following has been run:
# feats <- RBERT::extract_features(...) # See RBERT documentation
# Then:
attn <- format_attention(feats$attention_probs)
visualize_attention(attn)
}
}
